# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Serveur mock pour LilSnap",
    "description": "",
    "version": "1.0"
  },
  "produces": [
    "application/json"
  ],
...
  "paths": {
    "/history": {
      "get": {
        "summary": "Retourne l'historique des notifications, en excluant celles marquées comme supprimées.",
        "tags": [
          "Notifications"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "title": {
                        "type": "string"
                      },
                      "body": {
                        "type": "string"
                      },
                      "isAlreadyRead": {
                        "type": "boolean"
                      },
                      "isDeleted": {
                        "type": "boolean"
                      },
                      "date": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "type": {
                        "type": "string"
                      },
                      "urlExterne": {
                        "type": "string"
                      },
                      "urlActualite": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Envoie un tableau de notification et modifie les états lues et/ou supprimées des notifications du tableau.",
        "tags": [
          "Notifications"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "isAlreadyRead": {
                      "type": "boolean"
                    },
                    "isDeleted": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/terminal/{fcmoken}": {
      "put": {
        "summary": "Abonne un appareil à un topic Firebase Cloud Messaging.",
        "tags": [
          "Notifications"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "name": "fcmoken",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "delete": {
        "summary": "Désabonne un appareil d'un topic Firebase Cloud Messaging.",
        "tags": [
          "Notifications"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "name": "fcmoken",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/sendFCM/{fcmToken}": {
      "post": {
        "summary": "Envoie une notification à un appareil spécifique via un token d'enregistrement FCM (firebase).",
        "tags": [
          "Notifications"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "fcmToken",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "body": {
                    "type": "string"
                  },
                  "isAlreadyRead": {
                    "type": "boolean"
                  },
                  "isDeleted": {
                    "type": "boolean"
                  },
                  "date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "type": {
                    "type": "string"
                  },
                  "urlExterne": {
                    "type": "string"
                  },
                  "urlActualite": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sendTOPIC/{topic}": {
      "post": {
        "summary": "Envoie une notification à tous les appareils abonnés à un topic spécifique.",
        "tags": [
          "Notifications"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "topic",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "body": {
                    "type": "string"
                  },
                  "isAlreadyRead": {
                    "type": "boolean"
                  },
                  "isDeleted": {
                    "type": "boolean"
                  },
                  "date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "type": {
                    "type": "string"
                  },
                  "urlExterne": {
                    "type": "string"
                  },
                  "urlActualite": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/groupes/{groupeId}/abonnement": {
      "post": {
        "summary": "Ajoute un abonnement à un groupe spécifique par son ID.",
        "tags": [
          "Abonnements"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "name": "groupeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/abonnements/{groupeId}": {
      "delete": {
        "summary": "Supprime un abonnement d'un groupe spécifique par son ID.",
        "tags": [
          "Abonnements"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "name": "groupeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/semestres/{id}/cours": {
      "get": {
        "summary": "Retourne tous les cours associés à un semestre spécifique.",
        "tags": [
          "Abonnements"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "fatherId": {
                        "type": "integer"
                      },
                      "fatherName": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "id": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
    "/cours/{id}/groupes": {
      "get": {
        "summary": "Retourne tous les groupes associés à un cours spécifique.",
        "tags": [
          "Abonnements"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "fatherId": {
                        "type": "integer"
                      },
                      "fatherName": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "id": {
                        "type": "integer"
                      },
                      "type": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
...
    "/lieu/{ID}/image": {
      "get": {
        "summary": "Retourne l'image associée à un lieu spécifique par son ID.",
        "tags": [
          "Geolocalication"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "ID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  }
}